training_pipeline/requirements.txt
hydra-core>=1.3.0
omegaconf
pytorch-lightning>=2.0.0
torch==2.7.1
torchvision==0.22.1